backend/requirements.txt
fastapi
uvicorn 
sqlalchemy
python-multipart
python-dotenv
itsdangerous
jinja2
pillow
bleach